server/aws/requirements.txt
boto
boto3
pyyaml
paramiko
